+++ spec/truffle/interop/foreign_object_spec.rb
@@ -22,4 +22,5 @@ describe "Foreign objects" do
       p foreign
     }.should output_to_fd("#{foreign.inspect}\n")
   end
+
 end