+++ CHANGELOG.md
@@ -9,6 +9,7 @@ New features:
 * `foreign_object.inspect` has been improved to be more useful (include the language and meta object).
 * `foreign_object.class` now calls `getMetaObject()` (except for Java classes, same as before).
 * Add basic support for Linux ARM64.
+* `foreign_object.name = value` will now call `Interoplibrary#writeMember("name", value)` instead of `invokeMember("name=", value)`.
 
 Bug fixes: